Part 14. Spring Boot: JPA oneToMany - users messages, messages editor.

+++ src/main/java/ru/izotov/sweater/repo/MessageRepo.java
@@ -1,10 +1,10 @@
 package ru.izotov.sweater.repo;
 
-import org.springframework.data.repository.CrudRepository;
 import ru.izotov.sweater.domain.Message;
+import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTag(String tag);
 }
